requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
environ==1.0
idna==3.3
lxml==4.9.1
Markdown==3.4.1
Pillow==9.4.0
pytz==2022.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
